Upgrade PostgreSQL driver to 42.3.1

+++ plugin/trino-postgresql/src/main/java/io/trino/plugin/postgresql/TypeUtils.java
@@ -75,7 +75,7 @@ final class TypeUtils
     static String getArrayElementPgTypeName(ConnectorSession session, PostgreSqlClient client, Type elementType)
     {
         if (DOUBLE.equals(elementType)) {
-            return "float";
+            return "float8";
         }
 
         if (REAL.equals(elementType)) {
